+++ common/dataStructures/DiskSpaceReservationRequest.hpp
@@ -24,7 +24,6 @@ namespace cta {
 
 struct DiskSpaceReservationRequest: public std::map<std::string, uint64_t> {
   void addRequest(const std::string &diskSystemName, uint64_t size);
-  void removeRequest(const std::string &diskSystemName, uint64_t size);
 };
 
 }