+++ scheduler/RetrieveMount.hpp
@@ -212,6 +212,14 @@ namespace cta {
     virtual bool reserveDiskSpace(const cta::DiskSpaceReservationRequest &request, log::LogContext& logContext);
 
     /**
+     * Performs the disk space reservation logic for the request and mount, but does not reserve space in the catalogue
+     * @param request the disk space reservation request
+     * @param logContext
+     */
+    virtual bool testReserveDiskSpace(const cta::DiskSpaceReservationRequest &request, log::LogContext& logContext);
+
+
+    /**
      * Wait and complete reporting of a batch of jobs successes. The per jobs handling has
      * already been launched in the background using cta::RetrieveJob::asyncComplete().
      * This function will check completion of those async completes and then proceed