+++ scheduler/SchedulerDatabase.hpp
@@ -437,6 +437,9 @@ class SchedulerDatabase {
       uint64_t bytesRequested, log::LogContext& logContext) = 0;
     virtual bool reserveDiskSpace(const cta::DiskSpaceReservationRequest &request,
       const std::string &fetchEosFreeSpaceScript, log::LogContext& logContext) = 0;
+    virtual bool testReserveDiskSpace(const cta::DiskSpaceReservationRequest &request,
+      const std::string &fetchEosFreeSpaceScript, log::LogContext& logContext) = 0;
+
     virtual void requeueJobBatch(std::list<std::unique_ptr<SchedulerDatabase::RetrieveJob>>& jobBatch, 
       log::LogContext& logContext) = 0;